+++ packages/core/src/v3/runtime/noopRuntimeManager.ts
@@ -1,20 +1,9 @@
-import {
-  BatchTaskRunExecutionResult,
-  TaskMetadataWithFilePath,
-  TaskRunContext,
-  TaskRunExecutionResult,
-} from "../schemas";
+import { BatchTaskRunExecutionResult, TaskRunContext, TaskRunExecutionResult } from "../schemas";
 import { RuntimeManager } from "./manager";
 
 export class NoopRuntimeManager implements RuntimeManager {
   disable(): void {
     // do nothing
-  }
-
-  registerTasks(): void {}
-
-  getTaskMetadata(id: string): TaskMetadataWithFilePath | undefined {
-    return undefined;
   }
 
   waitForDuration(ms: number): Promise<void> {