+++ packages/database/prisma/migrations/20240411145517_added_trigger_source_to_background_worker_task/migration.sql
@@ -1,0 +1,5 @@
+-- CreateEnum
+CREATE TYPE "TaskTriggerSource" AS ENUM ('STANDARD', 'SCHEDULED');
+
+-- AlterTable
+ALTER TABLE "BackgroundWorkerTask" ADD COLUMN     "triggerSource" "TaskTriggerSource" NOT NULL DEFAULT 'STANDARD';