+++ packages/trigger-sdk/src/v3/index.ts
@@ -3,10 +3,24 @@ export * from "./config";
 export * from "./wait";
 export * from "./cache";
 export { retry, type RetryOptions } from "./retry";
+export { queue } from "./shared";
 
 import type { Context } from "./shared";
 export type { Context };
 
-export { logger, type LogLevel } from "@trigger.dev/core/v3";
+export {
+  logger,
+  type LogLevel,
+  APIError,
+  BadRequestError,
+  AuthenticationError,
+  PermissionDeniedError,
+  NotFoundError,
+  ConflictError,
+  UnprocessableEntityError,
+  RateLimitError,
+  InternalServerError,
+} from "@trigger.dev/core/v3";
 
 export { runs } from "./management";
+export * as schedules from "./schedules";