+++ references/v3-catalog/src/trigger/scheduled.ts
@@ -1,0 +1,11 @@
+import { logger, schedules } from "@trigger.dev/sdk/v3";
+
+export const firstScheduledTask = schedules.task({
+  id: "first-scheduled-task",
+  run: async (payload) => {
+    const distanceInMs =
+      payload.timestamp.getTime() - (payload.lastTimestamp ?? new Date()).getTime();
+
+    logger.log("First scheduled tasks", { payload, distanceInMs });
+  },
+});