+++ packages/nestjs/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @trigger.dev/nestjs
 
+## 3.0.0-beta.13
+
+### Patch Changes
+
+- Updated dependencies [4986bfda2]
+- Updated dependencies [44e1b8754]
+- Updated dependencies [4986bfda2]
+- Updated dependencies [4986bfda2]
+  - @trigger.dev/sdk@3.0.0-beta.13
+
 ## 3.0.0-beta.12
 
 ### Patch Changes